Edit the endpoint url
