+++ src/index.html
@@ -24,7 +24,7 @@
         <h2>Add your score</h2>
         <form action="#">
           <input type="text" placeholder="Your name" class="name">
-          <input type="text" placeholder="Your score" class="points">
+          <input type="text" placeholder="Your score" class="scores">
           <button type="submit" class="submit-btn">Submit</button>
         </form>
       </div>
